(get_args): --datacenter option removed. New options --network and --annotation. Add single-letter equivalents for all options.
(make_nic_changespec): New function.
(vmclone): Use it.
Do not set values for cluster, datastore, etc. if not specified by option.
The default is to inherit these from the vm/template being cloned.
The one value we still seem to need to set explicitly is the folder, but
get this from the template if not specified.

(myCallback.callback): Qualify printerr as vsl.printerr.