Fairly significant refactoring.
(cmds): Variable deleted.
(get_args): Pass loadrc=True to ArgumentParser.
Create a parser tree for subcommands which defines command names, options
specific to those commands, and help strings for everything.

(mVM): Every public method (i.e. those which implement commands) now take
only one fixed argument, usually vm_name.  There is one exception in kind
only for `register', which takes the path to a vmx file since the vm in
question is not yet in inventory.
(mVM.connect): This is the other slight exception; the optional `connect'
kwarg is there so that the `disconnect' method can call it with the
opposite value, since otherwise the two methods would be completely
identical.  It's not meant to be used publicly.

(main): Vastly simplify.  The only special thing it does now for any of the
commands is to follow up `register' commands with the `--keep-macs' option,
if specified.  Unfortunately there is no way to set that flag at the same
time as registration.