Major changes to pseudoPropAttr.
Objects have an attribute __vimtype__ recording type of the original object.
Automatically handle any x.key and x.value vim.DataObject object types.
For any object with a _GetPropertyList attribure, use that to fetch the
list of attributes to copy.

The class variable pseudoPropAttr.strict set to True will make references
to partial keys an exception.  For example, if the keys 'guestinfo.json'
and 'guestinfo.vmtools.description' are defined, accessing 'guestinfo' will
result in an exception since that's not one of the literal keys.  However,
accessing obj.guestinfo will return the pseudoPropAttr structure
regardless.

The method pseudoPropAttr.deep() will do a recursive conversion of the
argument and all subelements therein.  It will not convert
vim.ManagedObject types unless the keyword arg convertManagedObject is
true, but in any case it will not convert any ManagedObject subelements..
This is because a recursive traversal of such objects could result in
retrieval of every managed object in the vsphere inventory.

(pseudoPropAttr.deep): New method.

(get_seq_type): Check for __vimtype__ if direct object type does not match.